feat: add ask beacons to one-way swaps; relaxed two-way constraint
* Added ask beacons and changed the names for the offer beacons to
accomodate it. Moved the gen name functions out of the common
modules.
* Two-way swaps no longer require storing the swap with either
asset1 or asset2. The 2 ada deposit already disincentivizes
creating swaps for fake assets to denial-of-service attack the
protocol. This attack would only be relevant for offer and ask
queries since most users know exactly what conversion they want.
The offer and ask queries are more for arbitragers who will
likely have hardware capable of handling minor attacks.

Warning: this commit does not contain the updates to the CLI for
the new ask beacons.